+++ chemprop/cli/train.py
@@ -422,6 +422,7 @@ def add_train_args(parser: ArgumentParser) -> ArgumentParser:
     split_args.add_argument(
         "--save-smiles-splits",
         action="store_true",
+        default=False,
         help="Whether to store the SMILES in each train/val/test split",
     )
     split_args.add_argument(
@@ -987,6 +988,9 @@ def train_model(
         )
         trainer.fit(model, train_loader, val_loader)
 
+        if args.dry_run:
+            return
+
         if test_loader is not None:
             if isinstance(trainer.strategy, DDPStrategy):
                 torch.distributed.destroy_process_group()
@@ -1097,7 +1101,9 @@ def evaluate_and_save_predictions(preds, test_loader, metrics, model_output_dir,
         )
     else:
         df_preds = pd.DataFrame(list(zip(*namess, *preds.T)), columns=columns)
-    df_preds.to_csv(model_output_dir / "test_predictions.csv", index=False)
+
+    if not args.dry_run:
+        df_preds.to_csv(model_output_dir / "test_predictions.csv", index=False)
 
 
 def main(args):
@@ -1117,6 +1123,10 @@ def main(args):
     )
 
     splits = build_splits(args, format_kwargs, featurization_kwargs)
+
+    if args.dry_run:
+        temp_output_dir = TemporaryDirectory()
+        args.output_dir = Path(temp_output_dir.name)
 
     for fold_idx, (train_data, val_data, test_data) in enumerate(zip(*splits)):
         if args.num_folds == 1:
